requirements.txt
pyaudio
SpeechRecognition
PySerial
PyGame
regex
nltk